requirements/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.8
# by the following command:
#
#    pip-compile requirements.in
#
aiohttp==3.8.3
    # via -r requirements.in
aiosignal==1.3.1
    # via aiohttp
anyio==3.6.2
    # via jupyter-server
argon2-cffi==21.3.0
    # via jupyter-server
argon2-cffi-bindings==21.2.0
    # via argon2-cffi
asttokens==2.2.1
    # via stack-data
async-timeout==4.0.2
    # via aiohttp
attrs==22.2.0
    # via
    #   aiohttp
    #   jsonschema
babel==2.11.0
    # via jupyterlab-server
backcall==0.2.0
    # via ipython
beautifulsoup4==4.11.1
    # via nbconvert
bleach==5.0.1
    # via nbconvert
certifi==2022.12.7
    # via requests
cffi==1.15.1
    # via argon2-cffi-bindings
charset-normalizer==2.1.1
    # via
    #   aiohttp
    #   requests
comm==0.1.2
    # via ipykernel
contourpy==1.0.6
    # via matplotlib
cycler==0.11.0
    # via matplotlib
debugpy==1.6.5
    # via ipykernel
decorator==5.1.1
    # via ipython
defusedxml==0.7.1
    # via nbconvert
entrypoints==0.4
    # via jupyter-client
executing==1.2.0
    # via stack-data
fastjsonschema==2.16.2
    # via nbformat
fonttools==4.38.0
    # via matplotlib
frozenlist==1.3.3
    # via
    #   aiohttp
    #   aiosignal
idna==3.4
    # via
    #   anyio
    #   requests
    #   yarl
importlib-metadata==6.0.0
    # via
    #   jupyterlab-server
    #   nbconvert
importlib-resources==5.10.2
    # via jsonschema
ipykernel==6.19.4
    # via
    #   -r requirements.in
    #   ipywidgets
ipympl==0.9.2
    # via -r requirements.in
ipython==8.8.0
    # via
    #   ipykernel
    #   ipympl
    #   ipywidgets
ipython-genutils==0.2.0
    # via ipympl
ipyvue==1.8.0
    # via ipyvuetify
ipyvuetify==1.8.4
    # via -r requirements.in
ipywidgets==8.0.4
    # via
    #   -r requirements.in
    #   ipympl
    #   ipyvue
jedi==0.18.2
    # via ipython
jinja2==3.1.2
    # via
    #   jupyter-server
    #   jupyterlab-server
    #   nbconvert
json5==0.9.11
    # via jupyterlab-server
jsonschema==4.17.3
    # via
    #   jupyterlab-server
    #   nbformat
jupyter-client==7.4.1
    # via
    #   ipykernel
    #   jupyter-server
    #   nbclient
    #   voila
jupyter-core==5.1.2
    # via
    #   jupyter-client
    #   jupyter-server
    #   nbconvert
    #   nbformat
    #   voila
jupyter-server==1.23.4
    # via
    #   jupyterlab-server
    #   voila
jupyterlab-pygments==0.2.2
    # via nbconvert
jupyterlab-server==2.18.0
    # via voila
jupyterlab-widgets==3.0.5
    # via ipywidgets
jupytext==1.14.4
    # via -r requirements.in
kiwisolver==1.4.4
    # via matplotlib
lxml==4.9.2
    # via -r requirements.in
markdown-it-py==2.1.0
    # via
    #   jupytext
    #   mdit-py-plugins
markupsafe==2.1.1
    # via
    #   jinja2
    #   nbconvert
matplotlib==3.6.2
    # via
    #   -r requirements.in
    #   ipympl
matplotlib-inline==0.1.6
    # via
    #   ipykernel
    #   ipython
mdit-py-plugins==0.3.3
    # via jupytext
mdurl==0.1.2
    # via markdown-it-py
mistune==2.0.4
    # via nbconvert
multidict==6.0.4
    # via
    #   aiohttp
    #   yarl
nbclient==0.6.8
    # via
    #   -r requirements.in
    #   nbconvert
    #   voila
nbconvert==7.2.7
    # via
    #   jupyter-server
    #   voila
nbformat==5.7.1
    # via
    #   jupyter-server
    #   jupytext
    #   nbclient
    #   nbconvert
nest-asyncio==1.5.6
    # via
    #   ipykernel
    #   jupyter-client
    #   nbclient
numpy==1.24.1
    # via
    #   contourpy
    #   ipympl
    #   matplotlib
    #   pandas
packaging==22.0
    # via
    #   ipykernel
    #   jupyter-server
    #   jupyterlab-server
    #   matplotlib
    #   nbconvert
pandas==1.5.2
    # via -r requirements.in
pandocfilters==1.5.0
    # via nbconvert
parso==0.8.3
    # via jedi
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
pillow==9.4.0
    # via
    #   ipympl
    #   matplotlib
pkgutil-resolve-name==1.3.10
    # via jsonschema
platformdirs==2.6.2
    # via jupyter-core
prometheus-client==0.15.0
    # via jupyter-server
prompt-toolkit==3.0.36
    # via ipython
psutil==5.9.4
    # via ipykernel
ptyprocess==0.7.0
    # via
    #   pexpect
    #   terminado
pure-eval==0.2.2
    # via stack-data
pycparser==2.21
    # via cffi
pygments==2.14.0
    # via
    #   ipython
    #   nbconvert
pyparsing==3.0.9
    # via matplotlib
pyrsistent==0.19.3
    # via jsonschema
python-dateutil==2.8.2
    # via
    #   jupyter-client
    #   matplotlib
    #   pandas
python-dotenv==0.21.0
    # via -r requirements.in
pytz==2022.7
    # via
    #   babel
    #   pandas
pyyaml==6.0
    # via jupytext
pyzmq==24.0.1
    # via
    #   ipykernel
    #   jupyter-client
    #   jupyter-server
requests==2.28.1
    # via
    #   -r requirements.in
    #   jupyterlab-server
send2trash==1.8.0
    # via jupyter-server
six==1.16.0
    # via
    #   asttokens
    #   bleach
    #   python-dateutil
sniffio==1.3.0
    # via anyio
soupsieve==2.3.2.post1
    # via beautifulsoup4
stack-data==0.6.2
    # via ipython
terminado==0.17.1
    # via jupyter-server
tinycss2==1.2.1
    # via nbconvert
toml==0.10.2
    # via jupytext
tornado==6.2
    # via
    #   ipykernel
    #   jupyter-client
    #   jupyter-server
    #   terminado
traitlets==5.8.0
    # via
    #   comm
    #   ipykernel
    #   ipympl
    #   ipython
    #   ipywidgets
    #   jupyter-client
    #   jupyter-core
    #   jupyter-server
    #   matplotlib-inline
    #   nbclient
    #   nbconvert
    #   nbformat
    #   voila
urllib3==1.26.13
    # via requests
voila==0.5.0a0
    # via -r requirements.in
wcwidth==0.2.5
    # via prompt-toolkit
webencodings==0.5.1
    # via
    #   bleach
    #   tinycss2
websocket-client==1.4.2
    # via jupyter-server
websockets==10.4
    # via voila
widgetsnbextension==4.0.5
    # via ipywidgets
yarl==1.8.2
    # via aiohttp
zipp==3.11.0
    # via
    #   importlib-metadata
    #   importlib-resources
